+++ R/AIG.R
@@ -17,6 +17,10 @@
 #'
 #' The \code{\link{spatial3d_mirror}} creates the mirror image of the figure created using the spatial3d function.\cr
 #'
+#' The \code{\link{spatial3dDouble}} creates the mirror image of the figure created using the spatial3d function.\cr
+#'
+#' The \code{\link{spatial3d_mirrorDouble}} creates the variation of the images of the figure created using the spatial3d function based one four methods.\cr
+#'
 #' @references
 #' Sternberg, R. J. (1980). Representation and process in linear syllogistic reasoning. Journal of Experimental Psychology: General, 109(2), 119.
 #'