+++ R/MR3d2.R
@@ -23,7 +23,7 @@
 #' Bejar, I. I. (1990). A generative analysis of a three-dimensional spatial task. \emph{Applied Psychological Measurement}, 14(3), 237-245.
 #'
 #' @author Aiden Loe
-#' @title Spatial 3D Reasoning Item
+#' @title Spatial 3D Reasoning Item (double shapes)
 #' @seealso \code{\link{lisy}}, \code{\link{arith}}, \code{\link{spatial2d}}, \code{\link{spatial3d_mirror}}
 #' @return
 #' \describe{
@@ -31,11 +31,11 @@
 #' \item{rotationMatrix}{Return the cordinates of x,y,z,w}
 #'     }
 #' @examples
-#' item <- spatial3d(seed=4, angle=pi/1.3, x=0.3,y=4,z=0.8,axis = TRUE)
+#' item <- spatial3dDouble(seed=4, angle=pi/1.3, x=0.3,y=4,z=0.8,axis = TRUE)
 #'
 #' # To save the figure (not run)
 #' # library(rgl)
-#' # item <- spatial3d(seed=4, angle=pi/1.3, x=0.3,y=4,z=0.8,cubes=9,axis = TRUE)
+#' # item <- spatial3dDouble(seed=4, angle=pi/1.3, x=0.3,y=4,z=0.8,cubes=9,axis = TRUE)
 #'
 #' # save in pdf
 #' # wd<- '~/desktop'
@@ -47,31 +47,30 @@
 #' # save in png
 #' # rgl.snapshot(filename="image3D.png",fmt="png")
 
-spatial3d <- function(seed=1, angle=pi/1.3, x=0.3,y=3, z=0.8, cubes=8, axis = TRUE, filled='yes'){
-  #seed=1; angle=pi/1.3; x=0.3;y=3; z=0.8; cubes=8; axis = TRUE; filled='random'
+spatial3dDouble <- function(seed=1, angle=pi/1.3, x=0.3,y=3, z=0.8, cubes=8, axis = TRUE, filled='yes'){
+ # seed=1; angle=pi/1.3; x=0.3;y=3; z=0.8; cubes=8; axis = TRUE; filled='yes'
 
   # This will finalise the item
   set.seed(seed)
   clear3d()
   rgl.light()
-  vlist <- c(1,2,2)
 
-  #vlist <- c(6,2,2)
+  # first one
+  vlist <- c(1,2,2)
   (zlist <- matrix(NA,ncol=3,nrow=cubes))
   for (i in 1:cubes) { # 8 is the max. Arbitary number. Can be anything.
 
     if(filled=='random'){
-    sam <- sample(1:2,1)
-       if(sam==1){
+      sam <- sample(1:2,1)
+      if(sam==1){
         mycube <-  cube(vlist[1],vlist[2],vlist[3],filled=TRUE, fillcol='#4c4c4c')
-       }else{
-      mycube <- cube(vlist[1],vlist[2],vlist[3],filled=TRUE)
-       }
+      }else{
+        mycube <- cube(vlist[1],vlist[2],vlist[3],filled=TRUE)
+      }
     }
-
-     if(filled=='yes'){
-       cube(vlist[1],vlist[2],vlist[3],filled=TRUE)
-     }
+    if(filled=='yes'){
+      cube(vlist[1],vlist[2],vlist[3],filled=TRUE)
+    }
 
     if(filled=='no'){
       cube(vlist[1],vlist[2],vlist[3],filled=FALSE)
@@ -84,16 +83,44 @@ spatial3d <- function(seed=1, angle=pi/1.3, x=0.3,y=3, z=0.8, cubes=8, axis = TR
     zlist
   }
 
-  zlist
-  #clear3d()
-  #cube(zlist[1:cubes,1],zlist[1:cubes,2],zlist[1:cubes,3])
+  # second one
+  vlist <- c(8,2,2)
+  (xlist <- matrix(NA,ncol=3,nrow=cubes))
+  for (i in 1:cubes) { # 8 is the max. Arbitary number. Can be anything.
+
+    if(filled=='random'){
+      sam <- sample(1:2,1)
+      if(sam==1){
+         cube(vlist[1],vlist[2],vlist[3],filled=TRUE, fillcol='#4c4c4c')
+      }else{
+        cube(vlist[1],vlist[2],vlist[3],filled=TRUE)
+      }
+    }
+    if(filled=='yes'){
+      cube(vlist[1],vlist[2],vlist[3],filled=TRUE)
+    }
+
+    if(filled=='no'){
+      cube(vlist[1],vlist[2],vlist[3],filled=FALSE)
+    }
+
+
+    (step <- sample(1:3, 1))
+    (vlist[step] <- vlist[step]+1^rbinom(1,1,.25))
+    (xlist[i,] <- vlist)
+    xlist
+  }
+
+  (zxlist <- rbind(zlist, xlist))
+  clear3d()
+  cube(zxlist[1:16,1],zxlist[1:16,2],zxlist[1:16,3])
   rgl.viewpoint(fov = 0, userMatrix = rotationMatrix(angle=angle, x=x,y=y,z=z))
   roMatrix <- rotationMatrix(angle=angle, x=x,y=y,z=z)
   if(axis == TRUE){
-  axes3d( edges=c("x", "y", "z") )
-  title3d(main=NULL, sub=NULL, xlab='xlab', ylab='ylab', zlab='zlab')
+    axes3d( edges=c("x", "y", "z") )
+    title3d(main=NULL, sub=NULL, xlab='xlab', ylab='ylab', zlab='zlab')
   }
-  result <- list(figure=zlist, rotationMatrix = roMatrix)
+  result <- list(figure=zxlist, rotationMatrix = roMatrix)
   class(result) <- "threeD"
   return(result)
 }
@@ -108,12 +135,16 @@ spatial3d <- function(seed=1, angle=pi/1.3, x=0.3,y=3, z=0.8, cubes=8, axis = TR
 #' @param obj An object with class of threeD.
 #' @param angle,x,y,z  See details
 #' @param axis Showing the axis is helpful when first testing the function.
-#' @description This function generates the mirror image of the 3 dimensional display figure. It acts as a wrap because the creation of the figure is done using functions from the rgl package.
-#' @details For 3D figures, some of the cubes may be hidden in sight when automatically generated. Hence, one would need to rotate the display figure several times to ensure that none of the cubes are hidden.
+#' @param method There are 4 methods. See details.
+#' @param cubes The number of connected cubes together.
+#' @description This function generates the variation of the images of the 3 dimensional display figure. There are four methods in total. See details for more information. It acts as a wrap because the creation of the figure is done using functions from the rgl package.
+#' @details For 3D figures, some of the cubes may be hidden in sight when automatically generated. Hence, one would need to rotate the display figure several times to ensure that none of the cubes are hidden. This may be even more so when it comes to 2 figures within an image.
 #'
 #' The arguments angle, x, y, z represents the rotation of angle radians based on the x, y and z axis. This is a wrapper to the rotationMatrix function from the rgl package. Changing the values in the arguments angle, x, y, z coordinates allows one to programmatically change angles to study potential cognitive operators at work.
 #'
 #' You can also rotate the figure interactively by clicking on the figure and moving it in different direction.
+#'
+#' There are 4 methods to create distractors. The first method generates, the first image as a mirror, and the second image as normal. The second method is vice versa from the first method. The third method generates both mirror images. The first method generates correct rotation but the image is always shot of one cube.
 #'
 #' @references
 #' \url{https://en.wikipedia.org/wiki/Radian}
@@ -121,19 +152,20 @@ spatial3d <- function(seed=1, angle=pi/1.3, x=0.3,y=3, z=0.8, cubes=8, axis = TR
 #' Bejar, I. I. (1990). A generative analysis of a three-dimensional spatial task. \emph{Applied Psychological Measurement}, 14(3), 237-245.
 #'
 #' @author Aiden Loe
-#' @title Mirror Spatial 3D Reasoning Item
+#' @title Mirror Spatial 3D Reasoning Item (2)
 #' @seealso \code{\link{lisy}}, \code{\link{arith}}, \code{\link{spatial2d}}, \code{\link{spatial3d}}
 #' @return
 #' \describe{
 #' \item{figure}{Return the matrix that generates the mirror image of the display figure.}
 #'     }
 #' @examples
-#' display <- spatial3d(seed=4, angle=pi/1.3, x=0.3,y=4,z=0.8,axis = TRUE)
-#' display_mirror <- spatial3d_mirror(display, angle=pi/1.3, x=0.3,y=4,z=0.8,axis = TRUE)
+#' display <- spatial3dDouble(seed=4, angle=pi/1.3, x=0.3,y=4,z=0.8,axis = TRUE, cubes=8)
+#' display_mirror <- spatial3d_mirrorDouble(display, angle=pi/1.3, x=0.3,y=4,z=0.8,
+#'                                          axis = TRUE, method="one", cubes=8)
 #'
 #' # To save the figure (not run)
 #' # library(rgl)
-#' # display_mirror <- spatial3d_mirror(display, angle=pi/1.3, x=0.3,y=4,z=0.8,axis = TRUE)
+#' # display_mirror <- spatial3d_mirrorDouble(display, angle=pi/1.3, x=0.3,y=4,z=0.8,axis = TRUE)
 #' # wd<- '~/desktop'
 #'
 #' # save in pdf
@@ -145,14 +177,44 @@ spatial3d <- function(seed=1, angle=pi/1.3, x=0.3,y=3, z=0.8, cubes=8, axis = TR
 #' # rgl.snapshot(filename="image3D.png",fmt="png")
 #'
 
-spatial3d_mirror <- function(obj, angle=pi/1.3, x=0.3,y=3, z=0.8,axis = TRUE){
+spatial3d_mirrorDouble <- function(obj, angle=pi/1.3, x=0.3,y=3, z=0.8,axis = TRUE, method="one", cubes=8){
   result <- obj
   if(class(result)!="threeD"){
     stop("You need to use an object of class threeD.")
   }
+
+  # first shape mirror, second shape non mirror
+  if(method=="one"){
   clear3d()
   rgl.light()
-  cube(result$figure[1:nrow(result$figure),3],result$figure[1:nrow(result$figure),2],result$figure[1:nrow(result$figure),1]) #mirror
+  cube(result$figure[1:(nrow(result$figure)/2),3],result$figure[1:(nrow(result$figure)/2),2],result$figure[1:(nrow(result$figure)/2),1]) #mirror
+  cube(result$figure[cubes+1:cubes,1],result$figure[cubes+1:cubes,2],result$figure[cubes+1:cubes,3]) # non mirror
+  }
+
+  # first shape non mirror, second shape mirror
+  if(method=="two"){
+    clear3d()
+    rgl.light()
+    cube(result$figure[1:(nrow(result$figure)/2),1],result$figure[1:(nrow(result$figure)/2),2],result$figure[1:(nrow(result$figure)/2),3]) # non-mirror
+    cube(result$figure[cubes+1:cubes,3],result$figure[cubes+1:cubes,2],result$figure[cubes+1:cubes,1]) #mirror
+  }
+
+  # both shapes are mirror
+  if(method=="three"){
+    clear3d()
+    rgl.light()
+    cube(result$figure[1:nrow(result$figure),3],result$figure[1:nrow(result$figure),2],result$figure[1:nrow(result$figure),1]) #mirror
+  }
+
+  if(method=="four"){
+    clear3d()
+    rgl.light()
+    cube(result$figure[1:(nrow(result$figure)/2-1),1],result$figure[1:(nrow(result$figure)/2-1),2],result$figure[1:(nrow(result$figure)/2-1),3]) # non-mirror
+    cube(result$figure[cubes+2:cubes,1],result$figure[cubes+2:cubes,2],result$figure[cubes+2:cubes,3]) #mirror
+  }
+
+
+
   rgl.viewpoint(fov = 0, userMatrix = rotationMatrix(angle=angle, x=x,y=y,z=z))
   mirrorResult<- cbind(result$figure[1:8,3],result$figure[1:8,2],result$figure[1:8,1])
   if(axis == TRUE){
@@ -161,5 +223,3 @@ spatial3d_mirror <- function(obj, angle=pi/1.3, x=0.3,y=3, z=0.8,axis = TRUE){
   }
   return(mirrorResult)
 }
-
-