+++ R/MRtwoD.R
@@ -65,13 +65,6 @@ spatial2d <- function(items ,wd=NULL, view="2D",seed=NULL, degree= 360, ansDegre
     stop("Working Directory does not exist.")
   }
 
-
-# view <- 'top'
-# ansDegree <- 180
-#   wd<- "~/desktop"
-#   items <- 2
-#   degree<- 360
-
 result <- NULL
 for(item in 1:items){