+++ NEWS.md
@@ -80,7 +80,7 @@
 
 ### MINOR CHANGES
 
-* users can decide on how many connected cubes they want in their 3d items. 
+* users can decide on how many connected blocks they want in their 3d items to make up the cube. 
 
 * fixed bug in 2d spatial items.